Moved to_data_format to generic_utils.py.
