+++ docs/autogen.py
@@ -100,7 +100,6 @@ EXCLUDE = {
     'get',
     'set_image_dim_ordering',
     'normalize_data_format',
-    'to_data_format',
     'image_dim_ordering',
     'get_variable_shape',
 }
