+++ keras/backend/tensorflow_backend.py
@@ -20,7 +20,7 @@ import os
 from .common import floatx
 from .common import epsilon
 from .common import normalize_data_format
-from .common import to_data_format
+from ..utils.generic_utils import to_data_format
 from ..utils.generic_utils import has_arg
 
 # Legacy functions
