+++ keras/backend/theano_backend.py
@@ -23,7 +23,7 @@ import numpy as np
 from .common import floatx
 from .common import epsilon
 from .common import normalize_data_format
-from .common import to_data_format
+from ..utils.generic_utils import to_data_format
 from ..utils.generic_utils import has_arg
 # Legacy functions
 from .common import set_image_dim_ordering, image_dim_ordering
